+++ releases.md
@@ -1,5 +1,10 @@
 # Releases
 
+## Unreleased
+
+  - Use `String::Format` for formatting strings and restore {ruby Async::Job::Processor::Inline#status_string}.
+	- Add failed count to inline processor.
+
 ## v0.10.2
 
   - Minor code cleanup.